pyproject.toml
[project]
name = "scrapper"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "aiofiles>=24.1.0",
    "aiohttp>=3.11.13",
    "pillow>=11.1.0",
    "requests>=2.32.3",
    "tqdm>=4.67.1",
]

